feat: modularise event handlers
