+++ src/events/clientReady.ts
@@ -1,0 +1,18 @@
+import { ExtendedClient } from "../interfaces/ExtendedClient";
+import { errorHandler } from "../utils/errorHandler";
+import { logHandler } from "../utils/logHandler";
+import { registerCommands } from "../utils/registerCommands";
+
+/**
+ * Handles the ClientReady event from Discord.
+ *
+ * @param {ExtendedClient} bot The bot's Discord instance.
+ */
+export const clientReady = async (bot: ExtendedClient) => {
+  try {
+    await registerCommands(bot);
+    logHandler.log("info", "Bot is ready.");
+  } catch (err) {
+    await errorHandler(bot, "client ready event", err);
+  }
+};
